requirements.txt
anyio==4.9.0
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==4.4.2
executing==2.2.0
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
icecream==2.1.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
install-playwright==0.1.1
jinja2==3.1.6
llvmlite==0.44.0
lottie==0.7.1
markdown==3.8
markdownify==1.1.0
markupsafe==3.0.2
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
pillow==10.4.0
playwright==1.51.0
proglog==0.1.11
pydub==0.25.1
pyee==12.1.1
pygments==2.19.1
pylottie==2024
python-dotenv==1.1.0
python-telegram-bot==22.0
regex==2024.11.6
requests==2.32.3
setuptools==79.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing-extensions==4.13.2
urllib3==2.4.0
whisper==1.1.10
